refactor(manager): integrate domain events into manager execution
- Update command_builder to emit domain events
- Update flow_manager to use event-driven state transitions
- Update manager_executor to publish events at lifecycle points
- Update manager_run_coordinator for event-driven coordination
- Update manager_run_service to integrate event handlers
- Update prompts to include event emission points
- Update result_handler to emit completion events

Events integrated:
- ManagerExecutionStarted/Completed
- ManagerFlowDispatched/Queued
- IssueStateChanged events
- PlanCompleted/ReviewCompleted events
