+++ src/vibe3/manager/result_handler.py
@@ -11,8 +11,8 @@ from vibe3.models.orchestration import IssueState
 
 if TYPE_CHECKING:
     from vibe3.manager.flow_manager import FlowManager
+    from vibe3.models.orchestra_config import OrchestraConfig
     from vibe3.models.orchestration import IssueInfo
-    from vibe3.orchestra.config import OrchestraConfig
 
 
 class DispatchResultHandler:
